The image is finally wrapping itslef:
I set the width of the image to 600px and then height to auto, and then set the parent container to flex-wrap: wrap.
The Paragrapgh text was fitting it's text to max-content, then I set the main Paragrapgh's width to 600px, the same as the image.
And then finally the image was wrapping when the window was shrinked.